// src/internal/client/models/TemplatebackendUser.ts
export interface TemplatebackendUser {
    /**
     * 
     * @type {number}
     * @memberof TemplatebackendUser
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof TemplatebackendUser
     */
    tenantid?: number;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatebackendUser
     */
    status?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof TemplatebackendUser
     */
    roles?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof TemplatebackendUser
     */
    totpEnabled?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof TemplatebackendUser
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TemplatebackendUser
     */
    updatedAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof TemplatebackendUser
     */
    passwordChanged?: boolean;
}